pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["data", "models", "utils"]

[project]
name = "steganography-IP-protection"
version = "0.1.0"
description = "A system for hiding signed metadata in images"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cryptography==46.0.5",
    "numpy==2.4.3",
    "opencv-python==4.13.0.92",
    "psycopg2-binary==2.9.11", 
    "python-dotenv==1.2.2",
    "reedsolo==1.7.0",
    "torch==2.10.0",
    "torchvision==0.25.0",
    "kornia==0.8.2",
    "steganogan==0.1.2", # - used for downloading pretrained weights
    "lpips==0.1.4",
    "Pillow==10.2.0",
    "requests==2.33.1", # - used for dataset download
    "tqdm==4.67.3", # - used for dataset download
    "fastapi[standard]==0.135.3"
]

[project.optional-dependencies]
test = [
    "pytest==9.0.2",
    "pytest-mock==3.15.1"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]